whatsapp-ip/requirements.txt
# WhatsApp IP - Python Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# WhatsApp Cloud API
requests==2.31.0

# Google Workspace APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Database
SQLAlchemy==2.0.25

# AI/LLM (choose one or both)
openai==1.6.1
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# Security & Encryption
cryptography==41.0.7

# Utilities
python-dateutil==2.8.2

# Development Tools
ngrok==1.1.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
